misc changes

+++ projects/android/classygames/src/com/charlesmadere/android/classygames/models/games/Position.java
@@ -160,7 +160,11 @@ public final class Position
 	 */
 	public void removePiece()
 	{
-		piece.kill();
+		if (hasPiece())
+		{
+			piece.kill();
+		}
+
 		piece = null;
 	}
